Update config location in default flows
The previous code assumed using the library in a script running out of
the source tree as the current working directory. This change makes
the code expect to find these files wherever the python library is
located. This is necessary for loading these files from an installed
version of this package.

Signed-off-by: Russell Bryant <rbryant@redhat.com>