Add simple pipeline for use with default merlinite
The CLI's default model is quantized merlinite, and it does not seem
good enough to follow the instructions in the full pipeline included
in the new library.

This change attempts to still use the new library, but with a very
minimal configuration. It's not doing any validation on the output, so
the output is not going to be great.  Then again, the output has never
been great doing SDG with merlinite and the old sdg implementation.
This at least keeps the ability to a basic workflow test and demo on a
smaller system.

Signed-off-by: Russell Bryant <rbryant@redhat.com>