Add unit tests for sdg.filterblock
This new file introduces some unit tests for the filterblock module.
A previous commit introduced a bug fix to ensure that unexpected data
does not cause this block to break. These tests exercise those fixes.
The tests introduced here will run automatically in CI and help ensure
that the problem does not reoccur.

Signed-off-by: Russell Bryant <rbryant@redhat.com>