Make FlowParams available to blocks as PipelineContext
There is a lot of boilerplate in flows which simply passes values
from FlowParams down to a block.

Instead, lets put all of these parameters in a top-level context
object (PipelineContext) that blocks can reference, and remove
this stuff from block configs completely.

This makes flows substantially easier to read and reason about,
because they now contain more-or-less only the parameters that
are unique to that flow.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>